AU-2354: feature: AU-2354: Moved everything related to inputs from form_helpers to input_helpers.
