+++ e2e/utils/form_helpers.ts
@@ -2,17 +2,10 @@ import cloneDeep from "lodash.clonedeep"
 import {logger} from "./logger";
 import {hideSlidePopup} from "./helpers";
 import {validateFormErrors} from "./error_validation_helpers";
-import {Page, expect, Locator, test} from "@playwright/test";
-import {
-  FormData,
-  Selector,
-  PageHandlers,
-  FormFieldWithRemove,
-  isMultiValueField,
-  isDynamicMultiValueField, FormPage
-} from "./data/test_data"
-
 import {saveObjectToEnv, extractPath} from "./helpers";
+import {fillFormField, clickButton} from './input_helpers'
+import {Page, expect, test} from "@playwright/test";
+import {FormData, PageHandlers, FormPage} from "./data/test_data"
 
 /**
  * Fill form pages from given data array. Calls the pagehandler callbacks for
@@ -196,13 +189,8 @@ const fillProfileForm = async (
     const buttons = [];
     for (const [itemKey, itemField] of Object.entries(formPageObject.items)) {
       if (itemField.role === 'button') {
-        // Collect buttons to be clicked later.
         buttons.push(itemField);
-      } else if (itemField.role === 'multivalue') {
-        // Process multi-value fields separately.
-        await fillMultiValueField(page, itemField, itemKey);
       } else {
-        // Or fill simple form field.
         await fillFormField(page, itemField, itemKey);
       }
     }
@@ -324,570 +312,6 @@ const validateHiddenFields = async (page: Page, itemsToBeHidden: string[], formP
     await expect(page.locator(hiddenSelector), `Field ${hiddenItem} is not hidden on ${formPageKey}.`).not.toBeVisible();
     logger(`Field ${hiddenItem} is hidden on ${formPageKey}.`)
   }
-}
-
-/**
- * The fillDynamicMultiValueField function.
- *
- * This functions fills dynamic multi-value fields. Dynamic multi-value
- * fields have a radio button that toggles their visibility. After the radio
- * button has been pressed, the field is passed to fillMultiValueField where
- * data entry is performed.
- *
- * @param page
- *   Page object from Playwright.
- * @param formField
- *   The form field from the form data.
- * @param itemKey
- *   Element key in data definition.
- */
-const fillDynamicMultiValueField = async (page: Page, formField: Partial<FormFieldWithRemove>, itemKey: string) => {
-
-  // Check that we have the needed dynamic multi-value field.
-  if (!isDynamicMultiValueField(formField.dynamic_multi)) {
-    logger('A dynamic multi-value field has not been defined in:', itemKey);
-    return;
-  }
-
-  // Setup constants from the dynamic multi-value field.
-  const dynamicMultiValueField = formField.dynamic_multi;
-  const radioSelector = dynamicMultiValueField.radioSelector;
-  const revealedElementSelector = dynamicMultiValueField.revealedElementSelector;
-
-  // Click the radio button and wait for the multi-value field to appear.
-  await fillRadioField(radioSelector, itemKey, page);
-
-  // Fill the multi-value field.
-  if (revealedElementSelector.value) {
-    await page.locator(revealedElementSelector.value).waitFor({state: 'visible'}).then(async () => {
-      await fillMultiValueField(page, dynamicMultiValueField, itemKey);
-    });
-  } else {
-    logger('No revealed element selector defined for:', itemKey);
-  }
-}
-
-/**
- * The fillMultiValueField function.
- *
- * This function fills data to a multi-value field.
- * The fields in each multi-value item are iterated over,
- * updating the items index on each iteration.
- *
- * @param page
- *   Page object from Playwright.
- * @param formField
- *   The form field from the form data.
- * @param itemKey
- *   Element key in data definition.
- */
-const fillMultiValueField = async (page: Page, formField: Partial<FormFieldWithRemove>, itemKey: string) => {
-
-  // Check that the dynamic multi-value field has a multi-value field.
-  if (!isMultiValueField(formField.multi)) {
-    logger('A multi-value field has not been defined in:', itemKey);
-    return;
-  }
-
-  // Setup constants from the multi-value field.
-  const multiValueField = formField.multi;
-  const multiValueFieldButtonSelector = multiValueField.buttonSelector;
-
-  // Check if we have an initial item.
-  const initialItem = replacePlaceholder('0', '[INDEX]', multiValueFieldButtonSelector.resultValue);
-  const initialItemSelector = `[data-drupal-selector="${initialItem}"]`;
-  const initialItemExists = await page.$(initialItemSelector) !== null;
-
-  // Loop through each entry in the multi-value field.
-  for (const [index, multiItem] of Object.entries(multiValueField.items)) {
-
-    if (index === '0' && !initialItemExists) {
-      await clickButton(page, multiValueFieldButtonSelector);
-    }
-
-    if (index !== '0') {
-      await clickButton(page, multiValueFieldButtonSelector);
-    }
-
-    // Make sure we have an item to fill, whether an "Add more" button was clicked or not.
-    const resultItemKey = replacePlaceholder(index.toString(), '[INDEX]', multiValueFieldButtonSelector.resultValue);
-    const resultItemSelector = `[data-drupal-selector="${resultItemKey}"]`;
-
-    // Wait for the new element to be visible and fill it out.
-    await page.locator(resultItemSelector).waitFor({state: 'visible'}).then(async () => {
-
-      // Loop through each field item in each multi-value field entry.
-      for (const fieldItem of multiItem) {
-
-        // Update selectors for each field to match the current index.
-        if (fieldItem.selector) {
-          fieldItem.selector.value = replacePlaceholder(
-            index.toString(),
-            "[INDEX]",
-            fieldItem.selector?.value ?? ''
-          );
-          fieldItem.selector.resultValue = replacePlaceholder(
-            index.toString(),
-            "[INDEX]",
-            fieldItem.selector?.resultValue ?? ''
-          );
-        }
-
-        // Fill form field normally with replaced indexes.
-        await fillFormField(page, fieldItem, itemKey);
-      }
-    });
-  }
-}
-
-/**
- * Fill input field.
- *
- * Available selectors are
- * - data-drupal-selector
- * - role-label
- * - role
- * - label
- * - text
- *
- * Please see documentation for addtional info:
- * @see https://helsinkisolutionoffice.atlassian.net/wiki/spaces/KAN/pages/8481833123/Regressiotestit+WIP#fillInputField
- *
- * @param value
- *  Value inserted to a given field
- * @param selector
- *  Selector object. See test_data.ts for details.
- * @param page
- *  Page object from Playwright.
- * @param itemKey
- *  Item key used in data definition.
- */
-async function fillInputField(value: string, selector: Selector | undefined, page: Page, itemKey: string) {
-  const stringValue = value.toString();
-
-  // If selector is not present, we cannot use this function.
-  if (!selector) {
-    return;
-  }
-
-  /**
-   * Fills fields with given types/roles.
-   */
-  switch (selector.type) {
-    case "data-drupal-selector":
-      const customSelector = `[data-drupal-selector="${selector.value}"]`;
-      // Fill field with selector
-      await page.locator(customSelector).fill(value);
-      break;
-
-
-    case "data-drupal-selector-sequential":
-      const customSequentialSelector = `[data-drupal-selector="${selector.value}"]`;
-      const element = page.locator(customSequentialSelector);
-      await element.waitFor({state: 'visible'}).then(async () => {
-        await element.fill('');
-        await element.pressSequentially(value);
-      });
-      break;
-
-    /**
-     * Fill element with role & label selector.
-     */
-    case 'role-label':
-      if (selector.details && selector.details.role && selector.details.label) {
-
-        // @ts-ignore
-        await page.getByRole(selector.details.role, selector.details.options)
-          .getByLabel(selector.details.label).fill(stringValue);
-
-      }
-      break;
-
-    /**
-     * Fill field with role only selector.
-     */
-    case 'role':
-      if (selector.details && selector.details.role && selector.details.options) {
-
-        // @ts-ignore
-        await page.getByRole(selector.details.role,
-          selector.details.options).fill(stringValue);
-      } else {
-        logger(`Input: Role - incorrect settings -> ${itemKey}`);
-      }
-      break;
-
-    /**
-     * Fill field with Label selector. This isn't very good way to do this,
-     * because labels can change and are subjective to used language.
-     */
-    case 'label':
-      if (selector.details && selector.details.label) {
-        await page.getByLabel(selector.details.label).fill(stringValue);
-
-      }
-      break;
-    case 'text':
-      if (selector.details && selector.details.text) {
-        await page.getByText(selector.details.text, selector.details.options).fill(stringValue);
-      }
-
-      break;
-  }
-}
-
-/**
- * Fills select fields.
- *
- * @param selector
- * @param page
- * @param value
- */
-async function fillSelectField(selector: Selector | Partial<FormFieldWithRemove> | undefined, page: Page, value: string | undefined) {
-
-  if (!selector) {
-    return;
-  }
-
-  switch (selector.type) {
-
-    case 'dom-id-first':
-      if (typeof selector.value === 'string') {
-        await page.locator(selector.value).selectOption({index: 1});
-      }
-      break;
-    case 'by-label':
-      if (selector.value && value) {
-        const customSelector = `[data-drupal-selector="${selector.value}"]`;
-        await page.waitForSelector(customSelector);
-        await page.locator(customSelector).selectOption({label: value});
-      }
-      break;
-    case 'data-drupal-selector':
-      const customSelector = `[${selector.type}="${selector.value}"]`;
-
-      // Use page.$eval to set the value of input elements
-      await page.$eval(customSelector, (element, value) => {
-        // Cast the element to HTMLSelectElement
-        const selectElement = element as HTMLSelectElement;
-
-        let newValue = value ?? 'XX';
-
-        if (value === 'use-random-value' || value == '') {
-          // Fetch all options from the select element
-          const options = Array.from(selectElement.options);
-
-          // Check if there are any options
-          if (options.length > 0) {
-            // Randomly select an option
-            const randomIndex = Math.floor(Math.random() * options.length);
-            const selectedOption = options[randomIndex];
-            newValue = selectedOption.value;
-          }
-        }
-
-        // Set the selected value of the select element
-        selectElement.value = newValue;
-
-        // Trigger the 'change' event to simulate user interaction
-        const event = new Event('change', {bubbles: true});
-        selectElement.dispatchEvent(event);
-      }, value);
-
-      // TODO: see if waits can be removed
-      await page.waitForTimeout(2000);
-
-      break;
-
-  }
-}
-
-/**
- * Fill a checkbox.
- *
- * @param selector
- * @param itemKey
- * @param page
- */
-async function fillCheckboxField(selector: Selector | undefined, itemKey: string, page: Page) {
-
-  if (!selector) {
-    return;
-  }
-
-  switch (selector.type) {
-
-    case 'data-drupal-selector':
-
-      if (selector.name && selector.value) {
-        const customSelector = `[${selector.type}="${selector.value}"]`;
-
-        await page.locator(customSelector).click();
-      } else {
-        logger(`Checkbox -> settings missing -> ${itemKey}`);
-      }
-
-      break;
-
-    case 'text':
-
-      if (selector.details && selector.details.text) {
-        await page.getByText(selector.details.text, selector.details.options).click();
-      } else {
-        logger(`Checkbox -> settings missing -> ${itemKey}`);
-      }
-
-      break;
-
-    case 'label':
-      if (selector.details && selector.details.label) {
-        await page.getByLabel(selector.details.label).check();
-      } else {
-        logger(`Checkbox -> settings missing -> ${itemKey}`);
-      }
-      break;
-
-  }
-}
-
-/**
- * Fill radio field.
- *
- * @param selector
- * @param itemKey
- * @param page
- */
-async function fillRadioField(selector: Selector | undefined, itemKey: string, page: Page) {
-
-  if (!selector) {
-    return;
-  }
-
-  switch (selector.type) {
-
-    case 'text':
-
-      if (selector.details && selector.details.text) {
-        await page.getByText(selector.details.text, selector.details.options).click();
-      } else {
-        logger(`Radio: Text -> settings missing -> ${itemKey}`);
-      }
-
-      break;
-
-    case 'dom-id':
-      const radioSelector = selector.value; // Change this to the actual selector of your radio button
-      await page.waitForSelector(radioSelector ?? '');
-
-      try {
-        if (radioSelector != null) {
-          await page.click(radioSelector);
-        }
-      } catch (error) {
-        logger('Error during click:', error);
-      }
-
-      // Wait for the change event to be processed
-      // @ts-ignore
-      await page.waitForSelector(radioSelector);
-      break;
-
-    case 'dom-id-label':
-      const labelSelector = `.option.hds-radio-button__label[for="${selector.value}"]`;
-      // Wait for the label to exist
-      await page.waitForSelector(labelSelector);
-
-      // Click on the label element
-      await page.click(labelSelector);
-      break;
-
-    case 'partial-for-attribute':
-      const labelForSelector = `.option.hds-radio-button__label[for*="${selector.value}"]`;
-      await page.waitForSelector(labelForSelector);
-      try {
-        await page.click(labelForSelector);
-      } catch (error) {
-        logger(`Error clicking label with partial 'for' attribute: ${selector.value}`, error);
-      }
-      break;
-
-  }
-}
-
-/**
- * Fill form field.
- *
- * @param page
- * @param formField
- * @param itemKey
- */
-const fillFormField = async (page: Page, formField: Partial<FormFieldWithRemove>, itemKey: string) => {
-  const {selector, value, role} = formField;
-
-  /**
-   * Fill normal input field.
-   */
-  if (role === "input" && value) {
-    await fillInputField(value, selector, page, itemKey);
-  }
-
-  /**
-   * Upload file via method.
-   */
-  if (role === "fileupload") {
-    if (!selector || !selector.value) {
-      return;
-    }
-    await uploadFile(
-      page,
-      selector.value,
-      selector.resultValue ?? '',
-      value
-    )
-  }
-
-  /**
-   * Fill select field
-   */
-  if (role === "select") {
-    await fillSelectField(selector, page, value);
-  }
-
-  /**
-   * Radio
-   */
-  if (role === 'radio') {
-    await fillRadioField(selector, itemKey, page);
-  }
-
-  /**
-   * Checkbox
-   */
-  if (role === 'checkbox') {
-    await fillCheckboxField(selector, itemKey, page);
-  }
-
-  /**
-   * Multi-value field.
-   */
-  if (role === 'multivalue') {
-    await fillMultiValueField(page, formField, itemKey);
-  }
-
-  /**
-   * Dynamic multi-value field.
-   */
-  if (role === 'dynamicmultivalue') {
-    await fillDynamicMultiValueField(page, formField, itemKey);
-  }
-
-};
-
-/**
- * Helper function to replace string in given string.
- *
- * @param index
- * @param placeholder
- * @param value
- */
-const replacePlaceholder = (index: string, placeholder: string, value: string | undefined) => {
-  if (!value) {
-    return;
-  }
-  return value.replace(placeholder, index);
-}
-
-/**
- * Click button on the page. Wait for target page to load before progressing.
- *
- * @param page
- * @param buttonSelector
- */
-const clickButton = async (
-  page: Page,
-  buttonSelector: Selector
-) => {
-  let element: Locator | null = null;
-
-  switch (buttonSelector.type) {
-    case 'data-drupal-selector':
-      element = page.locator(`[${buttonSelector.name}="${buttonSelector.value}"]`);
-      break;
-
-    case 'add-more-button':
-      element = page.getByRole('button', {name: buttonSelector.value});
-      break;
-
-    case 'form-topnavi-link':
-      element = page.locator(`li[data-webform-page="${buttonSelector.value}"] .grants-stepper__step__circle_container`);
-      break;
-  }
-
-  if (element) {
-    await Promise.all([
-      element.waitFor({state: 'visible'}),
-      element.click(),
-    ]);
-  }
-};
-
-/**
- * Helper function to generate selectors based on rules. Deprecated?
- *
- * @param type
- * @param selectorValue
- */
-const buildSelector = (type: string, selectorValue: string) => {
-  if (type === 'data-drupal-selector') {
-    return `[${type}="${selectorValue}"]`;
-  }
-
-  if (type === 'locator') {
-    return selectorValue;
-  }
-
-  // Return as is as default
-  return selectorValue;
-}
-
-/**
- * Upload file.
- *
- * @param page
- * @param uploadSelector
- * @param fileLinkSelector
- * @param filePath
- */
-const uploadFile = async (
-  page: Page,
-  uploadSelector: string,
-  fileLinkSelector: string,
-  filePath: string | undefined,
-) => {
-
-  if (!filePath) {
-    logger('No file defined in', uploadSelector);
-    return;
-  }
-
-  // Setup locators for file input and result link.
-  const fileInput = page.locator(uploadSelector);
-  const resultLink = page.locator(fileLinkSelector);
-
-  // Create a promise for the file upload.
-  const postResponsePromise = page.waitForResponse(response =>
-    response.request().method() === "POST" && response.status() === 200
-  );
-
-  // Wait for all promises to fulfill.
-  await Promise.all([
-    fileInput.waitFor({ state: 'attached', timeout: 30000 }),
-    fileInput.setInputFiles(filePath),
-    postResponsePromise,
-    resultLink.waitFor({ state: 'visible', timeout: 30000 }),
-    expect(fileInput).toBeHidden(),
-  ]);
-
 }
 
 /**
@@ -1085,16 +509,9 @@ async function fillHakijanTiedotUnregisteredCommunity(formItems: any, page: Page
 }
 
 export {
-  fillProfileForm,
-  fillFormField,
-  clickButton,
-  uploadFile,
   createFormData,
-  hideSlidePopup,
+  fillProfileForm,
   fillGrantsFormPage,
-  fillSelectField,
-  fillInputField,
-  fillCheckboxField,
   fillHakijanTiedotRegisteredCommunity,
   fillHakijanTiedotPrivatePerson,
   fillHakijanTiedotUnregisteredCommunity,
